feat: users table updated

+++ src/app/api/userProfile/route.ts
@@ -1,43 +1,26 @@
-import { authOptions, prisma } from '@sahayeta/app/lib'
-import { getServerSession } from 'next-auth/next'
+import { prisma } from '@sahayeta/app/lib'
 import { NextRequest, NextResponse } from 'next/server'
+import { useServerSession } from '../../utils/useServerSession'
 
 export async function GET() {
+  const currentUser = await useServerSession()
   try {
-    const session = await getServerSession()
-
-    if (session?.user) {
-      const user = await prisma.user.findFirst({
-        where: { email: session?.user?.email },
-        select: {
-          id: true,
-          name: true,
-          username: true,
-          profileImage: true,
-          role: true,
-          orgName: true,
-          email: true,
-          emailVerified: true,
-          bio: true,
-          phoneNum: true,
-          address: true,
-          dob: true,
-          ctzImg: false,
-          balance: true,
-          campaign: true
-        }
-      })
-
-      if (user) {
-        return NextResponse.json(user)
-      } else {
-        return NextResponse.json({ message: 'User not found' }, { status: 404 })
-      }
-    } else {
+    if (!currentUser) {
       return NextResponse.json(
-        { message: 'Unauthorized access' },
-        { status: 401 }
+        { message: 'You must be logged in.' },
+        { status: 404 }
       )
+    }
+    const id = currentUser.id
+    const user = await prisma.user.findUnique({
+      where: { id }
+    })
+
+    if (user) {
+      delete user.password
+      return NextResponse.json(user)
+    } else {
+      return NextResponse.json({ message: 'User not found' }, { status: 404 })
     }
   } catch (error) {
     return NextResponse.json(
@@ -48,34 +31,24 @@ export async function GET() {
 }
 
 export async function PATCH(req: NextRequest) {
-  const session = await getServerSession(authOptions)
-  type User = {
-    name?: string
-    email?: string
-    image?: string
-    userDetails?: {
-      id: string
-    }
-  }
-
+  const currentUser = await useServerSession()
   try {
-    const body = await req.json()
-    if (!session) {
-      NextResponse.json({ message: 'You must be logged in.' }, { status: 404 })
-    } else {
-      const id = (session.user as User).userDetails?.id
-      const updateUser = await prisma.user.update({
-        where: { id },
-        data: body
-      })
-      if (!updateUser) {
-        return NextResponse.json(
-          { message: 'Method not found' },
-          { status: 404 }
-        )
-      }
-      return NextResponse.json(updateUser)
+    if (!currentUser) {
+      return NextResponse.json(
+        { message: 'You must be logged in.' },
+        { status: 404 }
+      )
     }
+    const id = currentUser.id
+    const body = await req.json()
+    const updateUser = await prisma.user.update({
+      where: { id },
+      data: body
+    })
+    if (!updateUser) {
+      return NextResponse.json({ message: 'Method not found' }, { status: 404 })
+    }
+    return NextResponse.json(updateUser)
   } catch (error) {
     return NextResponse.json(
       { message: 'Error updating user', error },
@@ -85,17 +58,15 @@ export async function PATCH(req: NextRequest) {
 }
 
 export async function DELETE(req: NextRequest) {
-  const session = await getServerSession(authOptions)
-  type User = {
-    name?: string
-    email?: string
-    image?: string
-    userDetails?: {
-      id: string
-    }
-  }
+  const currentUser = await useServerSession()
   try {
-    const id = (session.user as User).userDetails?.id
+    if (!currentUser) {
+      return NextResponse.json(
+        { message: 'You must be logged in.' },
+        { status: 404 }
+      )
+    }
+    const id = currentUser.id
     await prisma.user.delete({
       where: { id }
     })
